# openapi.yaml
# yaml-language-server: $schema=https://spec.openapis.org/oas/3.0/schema/2021-09-28

openapi: 3.0.3
info:
  title: Podcaster
  version: "1.0"
paths:
  /:
    get:
      operationId: Index
      responses:
        "200":
          description: OK
          content:
            text/html:
              schema:
                type: string
  /app.js:
    get:
      operationId: IndexFile
      responses:
        "200":
          description: OK
          content:
            text/javascript:
              schema:
                type: string
  /api/subscriptions:
    get:
      operationId: Subscriptions
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Subscription"
        "400":
          description: error

  /api/subscription:
    post:
      operationId: RegisterSubscription
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
              - url
              properties:
                url:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Subscription"
        "400":
          description: error
  
  /api/subscription/{id}/-/fetch:
    post:
      operationId: FetchSubscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: OK
        "404":
          description: error

  /api/subscription/{id}/{entryId}/open:
    post:
      operationId: OpenedEntry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: entryId
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: OK
        "404":
          description: error

  /api/subscription/{id}/{entryId}:
    delete:
      operationId: DeleteEntry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: entryId
          required: true
          schema:
            type: integer
      responses:
        "202":
          description: OK
        "404":
          description: error

  /api/subscription/{id}:
    get:
      operationId: GetEntries
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Entry"
        "400":
          description: error
        "404":
          description: not found
    delete:
      operationId: DeleteSubscription
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: OK
        "400":
          description: error
        "404":
          description: error
  

components:
  schemas:
    Entry:
      type: object
      required:
      - id
      - title
      - description
      - url
      - new
      - publishedAt
      properties:
        id:
          type: integer
          format: uint64
          x-go-name: ID
        title:
          type: string
        description:
          type: string
        url:
          type: string
        new:
          type: boolean
        publishedAt:
          type: string
    Subscription:
      type: object
      required:
      - id
      - title
      - url
      - new_entry_count
      properties:
        id:
          type: integer
          format: uint64
          x-go-name: ID
        title:
          type: string
        url:
          type: string
        new_entry_count:
          type: integer
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
security:
  - basicAuth: []